+++ clients/include/testing_geql2_geqlf.hpp
@@ -285,13 +285,13 @@ void testing_geql2_geqlf(Arguments& argus)
                                   rocblas_status_invalid_size);
 
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(1);
+            rocsolver_bench_inform(inform_invalid_size);
 
         return;
     }
 
     // memory size query is necessary
-    if(!USE_ROCBLAS_REALLOC_ON_DEMAND)
+    if(argus.mem_query || !USE_ROCBLAS_REALLOC_ON_DEMAND)
     {
         CHECK_ROCBLAS_ERROR(rocblas_start_device_memory_size_query(handle));
         if(BATCHED)
@@ -303,6 +303,12 @@ void testing_geql2_geqlf(Arguments& argus)
 
         size_t size;
         CHECK_ROCBLAS_ERROR(rocblas_stop_device_memory_size_query(handle, &size));
+        if(argus.mem_query)
+        {
+            rocsolver_bench_inform(inform_mem_query, size);
+            return;
+        }
+
         CHECK_ROCBLAS_ERROR(rocblas_set_device_memory_size(handle, size));
     }
 
@@ -326,7 +332,7 @@ void testing_geql2_geqlf(Arguments& argus)
                                                         lda, stA, dIpiv.data(), stP, bc),
                                   rocblas_status_success);
             if(argus.timing)
-                ROCSOLVER_BENCH_INFORM(0);
+                rocsolver_bench_inform(inform_quick_return);
 
             return;
         }
@@ -363,7 +369,7 @@ void testing_geql2_geqlf(Arguments& argus)
                                                         lda, stA, dIpiv.data(), stP, bc),
                                   rocblas_status_success);
             if(argus.timing)
-                ROCSOLVER_BENCH_INFORM(0);
+                rocsolver_bench_inform(inform_quick_return);
 
             return;
         }