+++ clients/include/testing_getf2_getrf_npvt.hpp
@@ -316,13 +316,13 @@ void testing_getf2_getrf_npvt(Arguments& argus)
                                   rocblas_status_invalid_size);
 
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(1);
+            rocsolver_bench_inform(inform_invalid_size);
 
         return;
     }
 
     // memory size query is necessary
-    if(!USE_ROCBLAS_REALLOC_ON_DEMAND)
+    if(argus.mem_query || !USE_ROCBLAS_REALLOC_ON_DEMAND)
     {
         CHECK_ROCBLAS_ERROR(rocblas_start_device_memory_size_query(handle));
         if(BATCHED)
@@ -335,6 +335,12 @@ void testing_getf2_getrf_npvt(Arguments& argus)
 
         size_t size;
         CHECK_ROCBLAS_ERROR(rocblas_stop_device_memory_size_query(handle, &size));
+        if(argus.mem_query)
+        {
+            rocsolver_bench_inform(inform_mem_query, size);
+            return;
+        }
+
         CHECK_ROCBLAS_ERROR(rocblas_set_device_memory_size(handle, size));
     }
 
@@ -359,7 +365,7 @@ void testing_getf2_getrf_npvt(Arguments& argus)
                                                              dA.data(), lda, stA, dinfo.data(), bc),
                                   rocblas_status_success);
             if(argus.timing)
-                ROCSOLVER_BENCH_INFORM(0);
+                rocsolver_bench_inform(inform_quick_return);
 
             return;
         }
@@ -398,7 +404,7 @@ void testing_getf2_getrf_npvt(Arguments& argus)
                                                              dA.data(), lda, stA, dinfo.data(), bc),
                                   rocblas_status_success);
             if(argus.timing)
-                ROCSOLVER_BENCH_INFORM(0);
+                rocsolver_bench_inform(inform_quick_return);
 
             return;
         }