+++ clients/include/testing_managed_malloc.hpp
@@ -214,13 +214,13 @@ void testing_managed_malloc(Arguments& argus)
                               rocblas_status_invalid_size);
 
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(1);
+            rocsolver_bench_inform(inform_invalid_size);
 
         return;
     }
 
     // memory size query is necessary
-    if(!USE_ROCBLAS_REALLOC_ON_DEMAND)
+    if(argus.mem_query || !USE_ROCBLAS_REALLOC_ON_DEMAND)
     {
         CHECK_ROCBLAS_ERROR(rocblas_start_device_memory_size_query(handle));
         CHECK_ALLOC_QUERY(rocsolver_labrd(handle, m, n, nb, (T*)nullptr, lda, (S*)nullptr,
@@ -229,6 +229,12 @@ void testing_managed_malloc(Arguments& argus)
 
         size_t size;
         CHECK_ROCBLAS_ERROR(rocblas_stop_device_memory_size_query(handle, &size));
+        if(argus.mem_query)
+        {
+            rocsolver_bench_inform(inform_mem_query, size);
+            return;
+        }
+
         CHECK_ROCBLAS_ERROR(rocblas_set_device_memory_size(handle, size));
     }
 
@@ -253,7 +259,7 @@ void testing_managed_malloc(Arguments& argus)
                                               dXRes, ldx, dYRes, ldy),
                               rocblas_status_success);
         if(argus.timing)
-            ROCSOLVER_BENCH_INFORM(0);
+            rocsolver_bench_inform(inform_quick_return);
 
         return;
     }