+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 - Generalized symmetric- and hermitian-definite eigensolvers using Divide and Conquer algorithm:
     - SYGVD (with batched and strided\_batched versions)
     - HEGVD (with batched and strided\_batched versions)
+- Added --mem_query option to rocsolver-bench, which will print the amount of device memory required
+  by a function.
 
 ### Optimizations